requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.3.6
gunicorn==20.1.0